use a bigger type for req_mem_size
The original uint16_t is enough for about 8 rings of 1024 entries.
While this is enough for current purposes it raises a question with
regard to silent 32 to 16 bit casting in operations.

To avoid this problem and to be future proof, use uint32_t for this
field in both firmware and host driver.

Sync with firmware build: b6fe2937582f8a4d9bf038b91f30d8b550034343

Firmware updates since last sync:
b6fe293 use a bigger type for req_mem_size

+++ host_driver/fsl_c2x0_crypto_layer.h
@@ -152,7 +152,7 @@ typedef struct crypto_c_hs_mem {
 			uint8_t num_of_rps;
 			uint8_t max_pri;
 			uint8_t num_of_fwresp_rings;
-			uint16_t req_mem_size;
+			uint32_t req_mem_size;
 			uint32_t drv_resp_ring;
 			uint32_t fw_resp_ring;
 			uint32_t s_cntrs;
